requirements.conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.11.26=h06a4308_0
expat=2.6.4=h6a678d5_0
filelock=3.13.1=py312h06a4308_0
fsspec=2024.12.0=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py312h06a4308_1
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libprotobuf=4.25.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=2.1.3=py312h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpmath=1.3.0=py312h06a4308_0
ncurses=6.4=h6a678d5_0
networkx=3.3=py312h06a4308_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openssl=3.0.15=h5eee18b_0
pandas=2.2.3=pypi_0
pip=24.2=py312h06a4308_0
python=3.12.8=h5148396_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.3.1=cpu_py312hece47aa_0
pytz=2024.2=pypi_0
readline=8.2=h5eee18b_0
setuptools=75.1.0=py312h06a4308_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
typing_extensions=4.12.2=py312h06a4308_0
tzdata=2024.2=pypi_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
